Restructured the SAN panel
