+++ src/features/panel/play/PlayPanel.js
@@ -2,9 +2,9 @@ import React from 'react';
 import { Grid } from '@mui/material';
 import History from 'features/panel/History';
 import OpeningTable from 'features/panel/OpeningTable';
-import SanTable from 'features/panel/SanTable';
-import SanButtons from 'features/panel/SanButtons';
-import SanButtonsDialogs from 'features/panel/SanButtonsDialogs';
+import SanTable from 'features/panel/san/SanTable';
+import SanButtons from 'features/panel/san/SanButtons';
+import SanButtonsDialogs from 'features/panel/san/SanButtonsDialogs';
 import StartedButtons from 'features/panel/play/StartedButtons';
 import StartedDialogs from 'features/panel/play/StartedDialogs';
 import FinishedButtons from 'features/panel/play/FinishedButtons';
