+++ src/features/panel/san/SanButtonsDialogs.js
